pyproject.toml
[project]
name = "feed-to-succeed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.43.5",
    "fastexcel>=0.20.2",
    "loguru>=0.7.3",
    "polars>=1.40.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "snowflake-connector-python>=4.4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP", "B", "S", "SIM", "RUF", "T20"]
